requirements.txt
PyQt6>=6.6
pdfplumber>=0.11
pyautogui>=0.9
opencv-python>=4.9
Pillow>=10.2
pygetwindow>=0.0.9
